+++ auth/auth_test.go
@@ -20,9 +20,7 @@ func (*Auth) TestLookupIdent(c *C) {
 }
 
 func (*Auth) TestIsBoard(c *C) {
-	conf := config.ServerConfigs{}
-	conf.Boards.Enabled = []string{"a"}
-	config.Set(conf)
+	config.SetBoards([]string{"a"})
 
 	// Board exists
 	c.Assert(IsBoard("a"), Equals, true)