+++ config/config.go
@@ -4,230 +4,141 @@
 package config
 
 import (
-	"encoding/json"
-	"io/ioutil"
-	"os"
-	"path/filepath"
 	"sync"
 	"time"
-
-	"github.com/DisposaBoy/JsonConfigReader"
-	"github.com/Soreil/mnemonics"
-	"github.com/bakape/meguca/util"
 )
 
 var (
-	// Overridable path for tests
-	configRoot = "config"
-
 	// Ensures no reads happen, while the configuration is reloading
-	mu sync.RWMutex
+	globalMu, boardConfigMu, boardsMu sync.RWMutex
 
 	// Contains currently loaded server configuration. Intialised to an empty
 	// struct pointer, so we don't have to explicitly initialise it in all our
 	// tests.
-	config = &ServerConfigs{}
+	global = &Configs{}
+
+	// All existing boards
+	boards []string
+
+	// Map of board IDs to their cofiguration structs
+	boardConfigs map[string]BoardConfigs
 
 	// JSON of client-accessable configuration
 	clientJSON []byte
 
-	// Hash of the config file. Used live updating configuration on the client
+	// Hash of the gloabal configs. Used for live reloading configuration on the
+	// client
 	hash string
 )
 
-// ServerConfigs stores the global configuration
-type ServerConfigs struct {
-	HTTP          HTTPConfigs
-	Rethinkdb     RethinkDBConfig
-	Boards        BoardConfig
-	Staff         StaffConfig
-	Images        ImageConfig
-	Posts         PostConfig
-	Recaptcha     RecaptchaConfig
-	Radio, Pyu    bool
-	InfoBanner    string
-	FeedbackEmail string
-	DefaultCSS    string
+// Configs stores the global configuration
+type Configs struct {
+	SSL              bool          `json:"-"`
+	TrustProxies     bool          `json:"-" gorethink:"trustProxies"`
+	Gzip             bool          `json:"-" gorethink:"gzip"`
+	Prune            bool          `json:"-" gorethink:"prune"`
+	Radio            bool          `json:"radio" gorethink:"radio"`
+	WebmAudio        bool          `json:"-" gorethink:"webmAudio"`
+	Hats             bool          `json:"hats" gorethink:"hats"`
+	MaxWidth         uint16        `json:"-" gorethink:"maxWidth"`
+	MaxHeight        uint16        `json:"-" gorethink:"maxHeight"`
+	MaxThreads       int           `json:"-" gorethink:"maxThreads"`
+	MaxBump          int           `json:"-" gorethink:"maxBump"`
+	JPEGQuality      int           `json:"-"`
+	PNGQuality       int           `json:"-"`
+	ThreadCooldown   int           `json:"threadCooldown" gorethink:"threadCooldown"`
+	MaxSubjectLength int           `json:"maxSubjectLength" gorethink:"maxSubjectLength"`
+	MaxSize          int64         `json:"-" gorethink:"maxSize"`
+	Origin           string        `json:"-" gorethink:"origin"`
+	SSLCert          string        `json:"-"`
+	SSLKey           string        `json:"-"`
+	Frontpage        string        `json:"-" gorethink:"frontpage"`
+	DefaultCSS       string        `json:"defaultCSS" gorethink:"defaultCSS"`
+	Salt             string        `json:"-" gorethink:"salt"`
+	ExcludeRegex     string        `json:"excludeRegex" gorethink:"excludeRegex"`
+	FeedbackEmail    string        `json:"-" gorethink:"feedbackEmail"`
+	Links            [][2]string   `json:"links" gorethink:"links"`
+	Spoilers         []uint8       `json:"spoilers" gorethink:"spoliers"`
+	SessionExpiry    time.Duration `json:"-" gorethink:"sessionExpiry"`
 }
 
-// HTTPConfigs stores HTTP server configuration
-type HTTPConfigs struct {
-	SSL, TrustProxies, Gzip bool
-	Addr, Origin, Cert, Key string
-	Frontpage               string
+// Defaults contains the default server configuration values
+var Defaults = Configs{
+	Origin:           "localhost:8000",
+	TrustProxies:     false,
+	Gzip:             true,
+	SSL:              false,
+	SSLCert:          "",
+	SSLKey:           "",
+	Frontpage:        "",
+	MaxThreads:       100,
+	MaxBump:          1000,
+	Links:            [][2]string{{"4chan", "http://www.4chan.org/"}},
+	Prune:            true,
+	WebmAudio:        true,
+	Hats:             false,
+	JPEGQuality:      90,
+	PNGQuality:       20,
+	MaxSize:          3145728,
+	MaxHeight:        6000,
+	MaxWidth:         6000,
+	Spoilers:         []uint8{0},
+	DefaultCSS:       "moe",
+	ThreadCooldown:   60,
+	MaxSubjectLength: 50,
+	Salt:             "LALALALALALALALALALALALALALALALALALALALA",
+	ExcludeRegex:     "/[\u2000-\u200f\u202a-\u202f\u205f-\u206f]+/g",
+	Radio:            false,
+	SessionExpiry:    30,
+	FeedbackEmail:    "admin@email.com",
 }
 
-// RethinkDBConfig stores address and datbase name to connect to
-type RethinkDBConfig struct {
-	Addr, Db string
-}
-
-// BoardConfig stores overall board configuration
-type BoardConfig struct {
-	Enabled []string
-	Boards  map[string]struct {
-		MaxThreads, MaxBump int
-		Title               string
-	}
-	Links [][2]string
-	Prune bool
-}
-
-// StaffConfig stores moderation staff related configuration
-type StaffConfig struct {
-	Classes       map[string]StaffClass
-	SessionExpiry time.Duration
-}
-
-// StaffClass contains properties of a single staff personel type
-type StaffClass struct {
-	Alias   string
-	Members map[string]string
-	Rights  map[string]bool
-}
-
-// ImageConfig stores file upload processing and thumbnailing configuration
-type ImageConfig struct {
-	WebmAudio   bool
-	Hats        bool
-	JpegQuality int
-	PngQuality  int
-	Max         struct {
-		Size          int64
-		Width, Height uint16
-	}
-	Spoilers []uint8
-}
-
-// PostConfig stores configuration related to creating posts
-type PostConfig struct {
-	ThreadCreationCooldown, MaxSubjectLength int
-	ReadOnly, SageEnabled, ForcedAnon        bool
-	Salt, ExcludeRegex                       string
-}
-
-// RecaptchaConfig stores the public and private key fot Google ReCaptcha
-// authentication
-type RecaptchaConfig struct {
-	Public, Private string
-}
-
-// A subset of ServerConfigs, that is exported as JSON to all clients
-type clientConfigs struct {
-	Boards struct {
-		Enabled []string `json:"enabled"`
-		Boards  map[string]struct {
-			Title string `json:"title"`
-		} `json:"boards"`
-		Default string      `json:"default"`
-		Psuedo  [][2]string `json:"psuedo"`
-		Links   [][2]string `json:"links"`
-	} `json:"boards"`
-	Lang struct {
-		Enabled []string `json:"enabled"`
-		Default string   `json:"default"`
-	} `json:"lang"`
-	Staff struct {
-		Classes map[string]struct {
-			Alias  string          `json:"alias"`
-			Rights map[string]bool `json:"rights"`
-		} `json:"classes"`
-	} `json:"staff"`
-	Images struct {
-		Spoilers []int `json:"spoilers"`
-		Hats     bool  `json:"hats"`
-	} `json:"images"`
-	Banners    []string `json:"banners"`
-	FAQ        []string `json:"FAQ"`
-	Eightball  []string `json:"eightball"`
-	Radio      bool     `json:"radio"`
-	IllyaDance bool     `json:"illiyaDance"`
-	DefaultCSS string   `json:"defaultCSS"`
-	InfoBanner string   `json:"infoBanner"`
-}
-
-// LoadConfig reads and parses the JSON config file and thread-safely loads it
-// into the server
-func LoadConfig() error {
-	var (
-		tempServer ServerConfigs
-		tempClient clientConfigs
-		path       = filepath.FromSlash(configRoot + "/config.json")
-	)
-
-	file, err := os.Open(path)
-	if err != nil {
-		return util.WrapError("Error reading configuration file", err)
-	}
-	defer file.Close()
-
-	// Strip comments
-	buf, err := ioutil.ReadAll(JsonConfigReader.New(file))
-	if err != nil {
-		return parseError(err)
-	}
-
-	if err := json.Unmarshal(buf, &tempServer); err != nil {
-		return parseError(err)
-	}
-	if err := json.Unmarshal(buf, &tempClient); err != nil {
-		return parseError(err)
-	}
-
-	tempJSON, err := json.Marshal(tempClient)
-	if err != nil {
-		return parseError(err)
-	}
-	tempHash, err := util.HashBuffer(buf)
-	if err != nil {
-		return parseError(err)
-	}
-
-	mu.Lock()
-	defer mu.Unlock()
-
-	config = &tempServer
-	clientJSON = tempJSON
-	hash = tempHash
-	if err := mnemonic.SetSalt(config.Posts.Salt); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func parseError(err error) error {
-	return util.WrapError("Error parsing configuration file", err)
-}
+// BoardConfigs stores overall board configuration
+type BoardConfigs struct{}
 
 // Get returns a pointer to the current server configuration struct. Callers
 // should not modify this struct.
-func Get() *ServerConfigs {
-	mu.RLock()
-	defer mu.RUnlock()
-	return config
+func Get() *Configs {
+	globalMu.RLock()
+	defer globalMu.RUnlock()
+	return global
+}
+
+// Set sets the internal configuration struct. To be used only in tests.
+func Set(c Configs) {
+	globalMu.Lock()
+	defer globalMu.Unlock()
+	global = &c
 }
 
 // GetClient returns punlic availability configuration JSON and a truncated
 // configuration MD5 hash
 func GetClient() ([]byte, string) {
-	mu.RLock()
-	defer mu.RUnlock()
+	globalMu.RLock()
+	defer globalMu.RUnlock()
 	return clientJSON, hash
-}
-
-// Set sets the internal configuration struct. To be used only in tests.
-func Set(c ServerConfigs) {
-	mu.Lock()
-	defer mu.Unlock()
-	config = &c
 }
 
 // SetClient sets the client configuration JSON and hash. To be used only in
 // tests.
 func SetClient(json []byte, cHash string) {
-	mu.Lock()
-	defer mu.Unlock()
+	globalMu.Lock()
+	defer globalMu.Unlock()
 	clientJSON = json
 	hash = cHash
+}
+
+// GetBoards returns all boards currently existing. Receivers must not modify
+// the slice.
+func GetBoards() []string {
+	boardsMu.RLock()
+	defer boardsMu.RUnlock()
+	return boards
+}
+
+// SetBoards updates the all boards currently existing set
+func SetBoards(b []string) {
+	boardsMu.Lock()
+	defer boardsMu.Unlock()
+	boards = b
 }