+++ db/dbInit_test.go
@@ -16,18 +16,16 @@ type DBInit struct{}
 var _ = Suite(&DBInit{})
 
 // All other functions that depend on the database
-type DBSuite struct {
-	dbName string
-}
+type DBSuite struct{}
 
 var _ = Suite(&DBSuite{})
 
 var testDBName string
 
 func (d *DBSuite) SetUpSuite(c *C) {
-	d.dbName = UniqueDBName()
-	c.Assert(Connect(""), IsNil)
-	c.Assert(InitDB(d.dbName), IsNil)
+	DBName = UniqueDBName()
+	c.Assert(Connect(), IsNil)
+	c.Assert(InitDB(), IsNil)
 }
 
 func (*DBSuite) SetUpTest(c *C) {
@@ -35,14 +33,11 @@ func (*DBSuite) SetUpTest(c *C) {
 	for _, table := range AllTables {
 		c.Assert(Write(r.Table(table).Delete()), IsNil)
 	}
-
-	conf := config.ServerConfigs{}
-	conf.Boards.Enabled = []string{"a"}
-	config.Set(conf)
+	config.SetBoards([]string{"a"})
 }
 
 func (d *DBSuite) TearDownSuite(c *C) {
-	c.Assert(r.DBDrop(d.dbName).Exec(RSession), IsNil)
+	c.Assert(r.DBDrop(DBName).Exec(RSession), IsNil)
 	c.Assert(RSession.Close(), IsNil)
 }
 
@@ -66,13 +61,9 @@ func (*DBSuite) TestVerifyVersion(c *C) {
 }
 
 func (*DBInit) TestLoadDB(c *C) {
-	conf := config.ServerConfigs{}
-	conf.Rethinkdb.Addr = "localhost:28015"
-	dbName := UniqueDBName()
-	conf.Rethinkdb.Db = dbName
-	config.Set(conf)
+	DBName = UniqueDBName()
 	defer func() {
-		c.Assert(Write(r.DBDrop(dbName)), IsNil)
+		c.Assert(Write(r.DBDrop(DBName)), IsNil)
 		c.Assert(RSession.Close(), IsNil)
 	}()
 	c.Assert(LoadDB(), IsNil)