+++ server/init.go
@@ -8,7 +8,6 @@ import (
 	"os"
 	"runtime"
 
-	"github.com/bakape/meguca/config"
 	"github.com/bakape/meguca/db"
 	"github.com/bakape/meguca/imager"
 	"github.com/bakape/meguca/templates"
@@ -86,7 +85,6 @@ func printUsage() {
 
 func startServer() {
 	fns := []func() error{
-		config.LoadConfig,
 		templates.Compile,
 		db.LoadDB,
 		imager.InitImager,