+++ server/websockets/admin.go
@@ -1,0 +1,3 @@
+// Board creation and configuration
+
+package websockets